+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## CHANGE LOG
 
+### v6.1.8
+
+2015-05-13
+
+- 上传增加putpolicy2
+
 ### v6.1.7
 
 2015-05-09
 
-- 上传策略增加 callbackHost、persistentPipeline, callbackFetchKey
+- 上传putpolicy2增加 callbackHost、persistentPipeline, callbackFetchKey
 - 增加fetch 函数
 - imageview -> imageview2
 
@@ -13,7 +19,7 @@
 
 2014-10-31
 
-- 上传策略增加fsizelimit, insertonly
+- 上传putpolicy2增加fsizelimit, insertonly
 
 
 ### v6.1.5